moved dominance frontier logic to analysis mod
