+++ src/ir/analysis/dom_tree.rs
@@ -1,9 +1,10 @@
+use super::super::block::Block;
 use super::super::func::Func;
 use super::super::block::BlockId;
 use std::collections::{HashSet, HashMap};
 
-pub fn compute_dom_tree(cfg: &mut Func) {
-    let entry_block = if let Some(block) = cfg.get_entry_block() {
+pub fn compute_dom_tree(func: &mut Func) {
+    let entry_block = if let Some(block) = func.get_entry_block() {
         block
     } else {
         return;
@@ -14,8 +15,8 @@ pub fn compute_dom_tree(cfg: &mut Func) {
     let mut dom_tree: HashMap<BlockId, Vec<BlockId>> = HashMap::new();
 
     while let Some(block_id) = work_list.pop() {
-        let block = &cfg[block_id];
-        let mut dominated_blocks = cfg.compute_dominated_blocks(block);
+        let block = &func[block_id];
+        let mut dominated_blocks = compute_dominated_blocks(func, block);
         dominated_blocks.remove(&block_id);
 
         for (_, blocks)  in dom_tree.iter_mut() {
@@ -38,3 +39,75 @@ pub fn compute_dom_tree(cfg: &mut Func) {
     todo!() // save the tree somewhere
     // cfg.dom_tree = dom_tree;
 }
+
+pub fn compute_reachable_blocks(func: &Func, block: &Block) -> HashSet<BlockId> {
+    let mut reachable_blocks = HashSet::from([block.get_id()]);
+    let mut work_list: Vec<usize> = block.get_successors().to_vec();
+
+    while let Some(id) = work_list.pop() {
+        if reachable_blocks.get(&id).is_some() {
+            continue;
+        }
+
+        reachable_blocks.insert(id);
+
+        let successor = &func[id];
+
+        work_list.append(&mut successor.get_successors().to_vec())
+    }
+
+    reachable_blocks
+}
+
+pub fn compute_dominated_blocks(func: &Func, seed_block: &Block) -> HashSet<BlockId> {
+    let mut dominated_blocks: HashSet<BlockId> = compute_reachable_blocks(func, seed_block).into_iter().collect();
+    let mut work_list: Vec<BlockId> = dominated_blocks.iter().map(|id| *id).collect();
+
+    while let Some(id) = work_list.pop() {
+        if id == seed_block.get_id() {
+            continue;
+        }
+
+        let block = &func[id];
+
+        if block.get_predecessors().iter().find(|id| dominated_blocks.get(&id).is_none()).is_some() {
+            dominated_blocks.remove(&id);
+
+            for id in block.get_successors().iter() {
+                if dominated_blocks.get(&id).is_some() {
+                    work_list.push(*id);
+                }
+            }
+        }
+    }
+
+    dominated_blocks
+}
+
+pub fn compute_dominance_frontier(func: &Func, seed_block: &Block) -> HashSet<BlockId> {
+    let mut dominance_frontier: HashSet<BlockId> = HashSet::new();
+    let dominated_blocks = compute_dominated_blocks(func, seed_block);
+
+    for id in dominated_blocks.iter() {
+        let block = &func[*id];
+
+        for successor_id in block.get_successors().iter() {
+            if dominated_blocks.get(successor_id).is_none() {
+                dominance_frontier.insert(*successor_id);
+            }
+        }
+    }
+
+    dominance_frontier
+}
+
+/*
+pub fn compute_unreachable_blocks(&self) -> HashSet<BlockID> {
+    let entry_block = self.get_entry_block();
+    let reachable_blocks = self.compute_reachable_blocks(entry_block);
+    let all_blocks = self.blocks.iter().map(|b| b.id).collect::<HashSet<BlockID>>();
+
+    all_blocks.difference(&reachable_blocks).map(|id| *id).collect()
+}
+*/
+
