+++ src/ir/analysis/mod.rs
@@ -7,3 +7,4 @@ mod dom_tree;
 pub use liveness_dfa::LivenessDFA;
 pub use dfa::DFA;
 pub use memory_ssa::MemoryAccess;
+pub use dom_tree::compute_dominance_frontier;
